[Optimization] change all value number unit from long to biginteger

+++ app/src/main/java/io/goldstone/blockchain/crypto/eos/eosram/EOSRamModel.kt
@@ -6,6 +6,7 @@ import io.goldstone.blockchain.crypto.eos.transaction.EOSAuthorization
 import io.goldstone.blockchain.crypto.eos.transaction.EOSTransactionInfo
 import io.goldstone.blockchain.crypto.multichain.CoinSymbol
 import java.io.Serializable
+import java.math.BigInteger
 
 /**
  * @author KaySaith
@@ -17,7 +18,7 @@ data class EOSRamModel(
 	val authorizations: List<EOSAuthorization>,
 	val payerName: String,
 	val receiverName: String,
-	val eosAmount: Long
+	val eosAmount: BigInteger
 ) : Serializable, EOSModel {
 	@Throws
 	override fun createObject(): String {
